pc-server/pyproject.toml
[project]
name = "touchpc-server"
version = "1.0.0"
description = "TouchPC Remote Control Server"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "pynput>=1.7.6",
    "zeroconf>=0.131.0",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "qrcode>=7.4.2",
]

[project.scripts]
touchpc-server = "remotecontrol.__main__:main"
